Update Layout with Sidebar
