+++ src/components/Navbar.jsx
@@ -1,30 +1,69 @@
 import * as React from "react";
 
-import AppBar from "@mui/material/AppBar";
-import Box from "@mui/material/Box";
-import CssBaseline from "@mui/material/CssBaseline";
-import Divider from "@mui/material/Divider";
-import Drawer from "@mui/material/Drawer";
-import IconButton from "@mui/material/IconButton";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListItemText from "@mui/material/ListItemText";
-import MenuIcon from "@mui/icons-material/Menu";
-import Toolbar from "@mui/material/Toolbar";
-import Typography from "@mui/material/Typography";
+import {
+  AppBar,
+  Avatar,
+  Badge,
+  Box,
+  Button,
+  CssBaseline,
+  Divider,
+  Drawer,
+  IconButton,
+  List,
+  ListItem,
+  ListItemButton,
+  ListItemIcon,
+  ListItemText,
+  Menu,
+  MenuItem,
+  Toolbar,
+  Typography,
+  useTheme,
+} from "@mui/material";
 
-import { useNavigate } from "react-router-dom";
+import MenuIcon from "@mui/icons-material/Menu";
+import LogoutIcon from "@mui/icons-material/Logout";
+
+import { Link, useNavigate } from "react-router-dom";
 import PropTypes from "prop-types";
 
 import { mainNavbarItems } from "../contexts/NavbarItems";
 import { ThemeProvider } from "@emotion/react";
 import theme from "../theme/color";
+import { styled } from "@mui/system";
 
 const drawerWidth = 240;
 
-export function ResponsiveDrawer(props) {
+const StyledBadge = styled(Badge)(({ theme }) => ({
+  "& .MuiBadge-badge": {
+    backgroundColor: "#44b700",
+    color: "#44b700",
+    boxShadow: `0 0 0 2px ${theme}`,
+    "&::after": {
+      position: "absolute",
+      top: 0,
+      left: 0,
+      width: "100%",
+      height: "100%",
+      borderRadius: "50%",
+      animation: "ripple 1.2s infinite ease-in-out",
+      border: "1px solid currentColor",
+      content: '""',
+    },
+  },
+  "@keyframes ripple": {
+    "0%": {
+      transform: "scale(.8)",
+      opacity: 1,
+    },
+    "100%": {
+      transform: "scale(2.4)",
+      opacity: 0,
+    },
+  },
+}));
+export function NavBar(props) {
   const { window, children } = props;
   const [mobileOpen, setMobileOpen] = React.useState(false);
   const [isClosing, setIsClosing] = React.useState(false);
@@ -46,9 +85,30 @@ export function ResponsiveDrawer(props) {
     }
   };
 
+  const [anchorEl, setAnchorEl] = React.useState(null);
+  const open = Boolean(anchorEl);
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget);
+  };
+  const handleReturn = () => {
+    navigate("/");
+  };
+  const handleClose = () => {
+    navigate("/profile");
+    setAnchorEl(null);
+  };
+  const handleSetting = () => {
+    navigate("/setting");
+    setAnchorEl(null);
+  };
+  const handleLogOut = () => {
+    navigate("/login");
+    setAnchorEl(null);
+  };
+
   const drawer = (
     <div>
-      <Toolbar sx={{ backgroundColor: "#FFC876" }} />
+      <Toolbar sx={{ backgroundColor: "#FFC876", minHeight: "70px" }} />
       <Divider />
       <List>
         {mainNavbarItems.map((item, index) => (
@@ -73,14 +133,15 @@ export function ResponsiveDrawer(props) {
 
   return (
     <ThemeProvider theme={theme}>
-      <Box sx={{ display: "flex" }}>
+      <Box sx={{ display: "flex", flexGrow: 1 }}>
         <CssBaseline />
         <AppBar
           position="fixed"
           elevation={0}
           sx={{
-            width: { sm: `calc(100% - ${drawerWidth}px)` },
+            // width: { sm: `calc(100% - ${drawerWidth}px)` },
             ml: { sm: `${drawerWidth}px` },
+            zIndex: 1300,
           }}
         >
           <Toolbar>
@@ -93,9 +154,64 @@ export function ResponsiveDrawer(props) {
             >
               <MenuIcon />
             </IconButton>
+            <IconButton sx={{ m: 1 }} onClick={handleReturn}>
+              <Avatar src="../assets/images/Fake-Jordan.png" />
+            </IconButton>
             <Typography variant="h6" noWrap component="div">
-              Responsive drawer
+              Sport Mou
             </Typography>
+
+            <Box
+              sx={{
+                display: "flex",
+                justifyContent: "flex-end",
+                alignItems: "center",
+                flexGrow: 1,
+              }}
+            >
+              <Link to={`/sell`}>
+                <Button variant="contained" color="error">
+                  Sell
+                </Button>
+              </Link>
+              <IconButton
+                aria-controls={open ? "basic-menu" : undefined}
+                aria-haspopup="true"
+                aria-expanded={open ? "true" : undefined}
+                onClick={handleClick}
+              >
+                <StyledBadge
+                  overlap="circular"
+                  anchorOrigin={{ vertical: "bottom", horizontal: "right" }}
+                  variant="dot"
+                >
+                  <Avatar src="../assets/images/Fake-Jordan.png" />
+                </StyledBadge>
+              </IconButton>
+
+              <Menu
+                id="basic-menu"
+                anchorEl={anchorEl}
+                open={open}
+                onClose={handleClose}
+                MenuListProps={{
+                  "aria-labelledby": "basic-button",
+                }}
+              >
+                <MenuItem onClick={handleClose}>
+                  <LogoutIcon />
+                  Account
+                </MenuItem>
+                <MenuItem onClick={handleSetting}>
+                  <LogoutIcon />
+                  Setting
+                </MenuItem>
+                <MenuItem onClick={handleLogOut}>
+                  <LogoutIcon />
+                  Logout
+                </MenuItem>
+              </Menu>
+            </Box>
           </Toolbar>
         </AppBar>
         <Box
@@ -122,6 +238,7 @@ export function ResponsiveDrawer(props) {
                 boxSizing: "border-box",
                 width: drawerWidth,
               },
+              zIndex: 1400,
             }}
           >
             {drawer}
@@ -156,7 +273,7 @@ export function ResponsiveDrawer(props) {
   );
 }
 
-ResponsiveDrawer.propTypes = {
+NavBar.propTypes = {
   /**
    * Injected by the documentation to work in an iframe.
    * Remove this when copying and pasting into your project.
