+++ src/pages/coaching/Coaching.jsx
@@ -1,5 +1,4 @@
 import { Box, Button } from "@mui/material";
-import { ResponsiveDrawer } from "../../components/Navbar";
 import { ThemeProvider } from "@emotion/react";
 import theme from "../../theme/color";
 import EventCard from "../../components/EventCard";
@@ -17,29 +16,21 @@ export function Coaching() {
 
   return (
     <ThemeProvider theme={theme}>
-      <ResponsiveDrawer>
-        <Box
-          sx={{ display: "flex", justifyContent: "flex-end", width: "100%" }}
-        >
-          <Button
-            onClick={handleCreateClick}
-            variant="outlined"
-            color="primary"
-          >
-            Create
-          </Button>
-        </Box>
-        <Box sx={{ mt: 2 }}>
-          <Grid container spacing={4}>
-            {eventIds.map((id) => (
-              <Grid key={id} item xs={12} sm={6} md={4}>
-                <EventCard id={id} />
-              </Grid>
-            ))}
-          </Grid>
-        </Box>
-        <Outlet />
-      </ResponsiveDrawer>
+      <Box sx={{ display: "flex", justifyContent: "flex-end", width: "100%" }}>
+        <Button onClick={handleCreateClick} variant="outlined" color="primary">
+          Create
+        </Button>
+      </Box>
+      <Box sx={{ mt: 2 }}>
+        <Grid container spacing={4}>
+          {eventIds.map((id) => (
+            <Grid key={id} item xs={12} sm={6} md={4}>
+              <EventCard id={id} />
+            </Grid>
+          ))}
+        </Grid>
+      </Box>
+      <Outlet />
     </ThemeProvider>
   );
 }
